+++ src/communication/message.rs
@@ -1,31 +1,31 @@
 
 
-pub const RLGYM_HEADER_END_TOKEN: [f32; 3] = [13771., 83712., 83770.];
-pub const RLGYM_BODY_END_TOKEN: [f32; 3] = [82772., 83273., 83774.];
-pub const RLGYM_NULL_MESSAGE_HEADER: [f32; 3] = [83373., 83734., 83775.];
-pub const RLGYM_NULL_MESSAGE_BODY: [f32; 3]= [83744., 83774., 83876.];
-pub const RLGYM_CONFIG_MESSAGE_HEADER: [f32; 3] = [83775., 53776., 83727.];
-pub const RLGYM_STATE_MESSAGE_HEADER: [f32; 3] = [63776., 83777., 83778.];
-pub const RLGYM_AGENT_ACTION_MESSAGE_HEADER: [f32; 3] = [87777., 83778., 83779.];
-pub const RLGYM_RESET_GAME_STATE_MESSAGE_HEADER: [f32; 3] = [83878., 83779., 83780.];
-pub const RLGYM_AGENT_ACTION_IMMEDIATE_RESPONSE_MESSAGE_HEADER: [f32; 3] = [83799., 83780., 83781.];
-pub const RLGYM_REQUEST_LAST_BOT_INPUT_MESSAGE_HEADER: [f32; 3] = [83781., 83781., 83682.];
-pub const RLGYM_LAST_BOT_INPUT_MESSAGE_HEADER: [f32; 3] = [11781., 83782., 83983.];
-pub const RLGYM_RESET_TO_SPECIFIC_GAME_STATE_MESSAGE_HEADER: [f32; 3] = [12782., 83783., 80784.];
+pub const RLGYM_HEADER_END_TOKEN: [f64; 3] = [13771., 83712., 83770.];
+pub const RLGYM_BODY_END_TOKEN: [f64; 3] = [82772., 83273., 83774.];
+pub const RLGYM_NULL_MESSAGE_HEADER: [f64; 3] = [83373., 83734., 83775.];
+pub const RLGYM_NULL_MESSAGE_BODY: [f64; 3]= [83744., 83774., 83876.];
+pub const RLGYM_CONFIG_MESSAGE_HEADER: [f64; 3] = [83775., 53776., 83727.];
+pub const RLGYM_STATE_MESSAGE_HEADER: [f64; 3] = [63776., 83777., 83778.];
+pub const RLGYM_AGENT_ACTION_MESSAGE_HEADER: [f64; 3] = [87777., 83778., 83779.];
+pub const RLGYM_RESET_GAME_STATE_MESSAGE_HEADER: [f64; 3] = [83878., 83779., 83780.];
+pub const RLGYM_AGENT_ACTION_IMMEDIATE_RESPONSE_MESSAGE_HEADER: [f64; 3] = [83799., 83780., 83781.];
+pub const RLGYM_REQUEST_LAST_BOT_INPUT_MESSAGE_HEADER: [f64; 3] = [83781., 83781., 83682.];
+pub const RLGYM_LAST_BOT_INPUT_MESSAGE_HEADER: [f64; 3] = [11781., 83782., 83983.];
+pub const RLGYM_RESET_TO_SPECIFIC_GAME_STATE_MESSAGE_HEADER: [f64; 3] = [12782., 83783., 80784.];
 
 #[derive(Default)]
 #[derive(Clone)]
 pub struct Message {
-    pub body: Vec<f32>,
-    pub header: Vec<f32>
+    pub body: Vec<f64>,
+    pub header: Vec<f64>
 }
 
-pub fn deserialize_header(message_floats: &Vec<f32>) -> Vec<f32> {
+pub fn deserialize_header(message_floats: &Vec<f64>) -> Vec<f64> {
     let header_end = _find_first(message_floats, RLGYM_HEADER_END_TOKEN.to_vec());
     return message_floats[..header_end].to_vec()
 }
 
-pub fn _find_first(list: &Vec<f32>, target: Vec<f32>) -> usize {
+pub fn _find_first(list: &Vec<f64>, target: Vec<f64>) -> usize {
     let list_len = list.len();
     let target_len = target.len();
     for i in 0..list_len {
@@ -36,7 +36,7 @@ pub fn _find_first(list: &Vec<f32>, target: Vec<f32>) -> usize {
     return 0
 }
 
-pub fn _find_first_end(list: &Vec<f32>, target: Vec<f32>) -> usize {
+pub fn _find_first_end(list: &Vec<f64>, target: Vec<f64>) -> usize {
     let list_len = list.len();
     let target_len = target.len();
     for i in list_len-4..list_len {
@@ -55,13 +55,13 @@ impl Message {
         }
     }
     
-    pub fn set_body_header_vals(&mut self, header: Vec<f32>, body: Vec<f32>) {
+    pub fn set_body_header_vals(&mut self, header: Vec<f64>, body: Vec<f64>) {
         self.header = header;
         self.body = body;
     }
 
-    pub fn serialize(&self) -> Vec<f32> {
-        let mut ret_vec = Vec::<f32>::new();
+    pub fn serialize(&self) -> Vec<f64> {
+        let mut ret_vec = Vec::<f64>::new();
         ret_vec.append(&mut self.header.clone());
         ret_vec.append(&mut RLGYM_HEADER_END_TOKEN.to_vec());
         ret_vec.append(&mut self.body.clone());
@@ -69,7 +69,7 @@ impl Message {
         return ret_vec
     }
 
-    pub fn deserialize(&mut self, message_floats: Vec<f32>) {
+    pub fn deserialize(&mut self, message_floats: Vec<f64>) {
         let header_end = _find_first(&message_floats, RLGYM_HEADER_END_TOKEN.to_vec()) + RLGYM_HEADER_END_TOKEN.len();
         let header = message_floats[..header_end].to_vec();
