+++ src/gamestates/physics_object.rs
@@ -4,12 +4,12 @@ use crate::math::*;
 #[derive(Default)]
 #[derive(Clone)]
 pub struct PhysicsObject {
-    pub position: Vec<f32>,
-    pub quaternion: Vec<f32>,
-    pub linear_velocity: Vec<f32>,
-    pub angular_velocity: Vec<f32>,
-    pub euler_angles: Vec<f32>,
-    pub rotation_mtx: Array2<f32>,
+    pub position: Vec<f64>,
+    pub quaternion: Vec<f64>,
+    pub linear_velocity: Vec<f64>,
+    pub angular_velocity: Vec<f64>,
+    pub euler_angles: Vec<f64>,
+    pub rotation_mtx: Array2<f64>,
     pub has_computed_rot_mtx: bool,
     pub has_computed_euler_angles: bool
 }
@@ -22,7 +22,7 @@ impl PhysicsObject {
             linear_velocity: vec![0.; 3],
             angular_velocity: vec![0.; 3],
             euler_angles: vec![0.; 3],
-            rotation_mtx: Array2::<f32>::zeros((3, 3)),
+            rotation_mtx: Array2::<f64>::zeros((3, 3)),
             has_computed_euler_angles: false,
             has_computed_rot_mtx: false
         }
@@ -42,20 +42,20 @@ impl PhysicsObject {
     //     }
     // }
 
-    pub fn decode_car_data(&mut self, car_data: Vec<f32>) {
+    pub fn decode_car_data(&mut self, car_data: Vec<f64>) {
         self.position = car_data[..3].to_vec();
         self.quaternion = car_data[3..7].to_vec();
         self.linear_velocity = car_data[7..10].to_vec();
         self.angular_velocity = car_data[10..].to_vec();
     }
 
-    pub fn decode_ball_data(&mut self, ball_data: Vec<f32>) {
+    pub fn decode_ball_data(&mut self, ball_data: Vec<f64>) {
         self.position = ball_data[..3].to_vec();
         self.linear_velocity = ball_data[3..6].to_vec();
         self.angular_velocity = ball_data[6..9].to_vec();
     }
 
-    pub fn forward(&mut self) -> Vec<f32> {
+    pub fn forward(&mut self) -> Vec<f64> {
         let arr = &self.rotation_mtx();
         let partial_arr = arr.column(0);
         // [:, 0]
@@ -63,38 +63,38 @@ impl PhysicsObject {
         return partial_arr.to_owned().to_vec()
     }
 
-    pub fn right(&mut self) -> Vec<f32> {
+    pub fn right(&mut self) -> Vec<f64> {
         let arr = self.rotation_mtx();
         let partial_arr = arr.column(1);
         return partial_arr.to_owned().to_vec()
     }
 
-    pub fn left(&mut self) -> Vec<f32> {
+    pub fn left(&mut self) -> Vec<f64> {
         let arr = self.rotation_mtx();
         let partial_arr = arr.column(1);
         let res_arr = partial_arr.to_owned() * -1.;
         return res_arr.to_vec()
     }
 
-    pub fn up(&mut self) -> Vec<f32> {
+    pub fn up(&mut self) -> Vec<f64> {
         let arr = self.rotation_mtx();
         let partial_arr = arr.column(2);
         return partial_arr.to_owned().to_vec()
     }
 
-    pub fn pitch(&mut self) -> f32 {
+    pub fn pitch(&mut self) -> f64 {
         self.euler_angles()[0]
     }
 
-    pub fn yaw(&mut self) -> f32 {
+    pub fn yaw(&mut self) -> f64 {
         self.euler_angles()[1]
     }
 
-    pub fn roll(&mut self) -> f32 {
+    pub fn roll(&mut self) -> f64 {
         self.euler_angles()[2]
     }
 
-    pub fn euler_angles(&mut self) -> Vec<f32> {
+    pub fn euler_angles(&mut self) -> Vec<f64> {
         if !self.has_computed_euler_angles {
             self.euler_angles = quat_to_euler(&self.quaternion.to_vec());
             self.has_computed_euler_angles = true;
@@ -102,7 +102,7 @@ impl PhysicsObject {
         return self.euler_angles.clone()
     }
     
-    pub fn rotation_mtx(&mut self) -> Array2<f32> {
+    pub fn rotation_mtx(&mut self) -> Array2<f64> {
         if !self.has_computed_rot_mtx {
             self.rotation_mtx = quat_to_rot_mtx(&self.quaternion.to_vec());
             self.has_computed_rot_mtx = true;
@@ -111,8 +111,8 @@ impl PhysicsObject {
         return self.rotation_mtx.clone()
     }
 
-    pub fn serialize(&self) -> Vec<f32> {
-        let mut repr = Vec::<f32>::new();
+    pub fn serialize(&self) -> Vec<f64> {
+        let mut repr = Vec::<f64>::new();
 
         // repr.extend([&mut self.position, &mut self.quaternion.clone()]);
 
@@ -122,7 +122,7 @@ impl PhysicsObject {
         repr.append(&mut self.angular_velocity.clone());
         repr.append(&mut self.euler_angles.clone());
         
-        let mut row_vec = Vec::<f32>::new();
+        let mut row_vec = Vec::<f64>::new();
         for i in self.rotation_mtx.clone() {
             row_vec.push(i)
         }
